// reEnergyUI/src/app/pages/model/Assessor.model.ts
export class QualificationItem {
  id: number
  qualification: string
  details: string
}
export class StateItem {
  id: number
  statename: string
  statecode: string
  gstcode: string

}

export class AssessorSearchResponse {
  total: number;
  rowCount: number;
  current: number;
  rows = new Array<AssessorCls>();
}

export class AssessorCls {
  userid: number;
  asrname: string;
  email: string;
  mobileno1: string;
  mobileno2: string;

  dob: Date;
  dob_struct: DateStruct
  dob_day: number;
  dob_month: number;
  dob_year: number;

  aadharnumber: string;
  qualificationid: number;
  otherqualification: string;
  specialitiesids: string; // coma separated ids
  totalworkexp: number;
  currentorg: string;
  designation: string;
  orgaddress: string;
  stateid: number;
  districtid: number;
  statename: string;
  districtname: string;
  city: string;
  residentialaddress: string;
  pin: string;
  latitude: number;
  longitude: number;

  bankname: string;
  bankbranchname: string;
  bankaccountnumber: number;
  accountholdername: string;
  ifsc_code: string;
  pannumber: string;

  photourl: string;
  createdby: number;
  createdon: Date;
  isactive: boolean;  

  capacity: number;
  capacityname: string;
  is_profile_submitted:boolean
}


export class DateStruct {
  year: number;
  month: number;
  day: number
}

export class EntityOperationResult {
  isSuccess: boolean;
  id: number;
  good: string;
  bad: string;
  error: string;
  message: string;
}
export class CalendarInfo {
  message: string;
  item: {
    id: number;
    assessorid: number;
    title: string;
    start: Date;
  }
}
export class EditCalendarInfo {
  id: string | number;
  assessorid: number;
  title: string;
  start: Date;
  multipartdate: Date;

}